Change to tick based timing
Also theoretically support tempo changes on the fly, not tested yet

+++ src/midi_file_reader.h
@@ -96,8 +96,8 @@ public:
 
     double getEndTime();
         
-    float ticksPerBeat; // precision (number of ticks distinguishable per second)
-    float startingTempo;
+    int ticksPerBeat;
+    int startingTempo;
     
     bool useAbsoluteTicks = false;
